# docs/openapi.yaml
components:
  schemas:
...
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - /api/bigshop/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
    Ingredient:
      additionalProperties: false
      properties:
        baseUnit:
          type: string
        department:
          type: string
        displayUnit:
          type: string
        name:
          type: string
        pantryStaple:
          type: boolean
        quantity:
          type: string
        unit:
          type: string
        unitSizes:
          additionalProperties:
            format: double
            type: number
          type: object
      required:
        - name
        - unit
        - quantity
      type: object
...
    Recipe:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - /api/bigshop/schemas/Recipe.json
          format: uri
          readOnly: true
          type: string
        id:
          format: int64
          type: integer
        ingredients:
          items:
            $ref: "#/components/schemas/Ingredient"
          type:
            - array
            - "null"
        method:
          type: string
        name:
          type: string
        notes:
          type: string
        remoteUrl:
          type: string
        tags:
          items:
            type: string
          type:
            - array
            - "null"
      required:
        - name
        - remoteUrl
        - notes
        - method
        - ingredients
        - tags
      type: object
    RecipeSummary:
      additionalProperties: false
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        tags:
          items:
            type: string
          type:
            - array
            - "null"
      required:
        - name
        - id
        - tags
      type: object
...
info:
  description: The Go API backing Big Shop, a recipe management and meal planning app.
  title: Big Shop API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /recipe/{id}:
    get:
      description: Looks a Recipe up by numeric ID if `id` parses as an integer, otherwise by its URL slug.
      operationId: get-recipe
      parameters:
        - description: Numeric Recipe ID or URL slug
          in: path
          name: id
          required: true
          schema:
            description: Numeric Recipe ID or URL slug
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get a recipe by ID or slug
      tags:
        - Recipes
  /recipes:
    get:
      description: Returns a lightweight (name/id/tags only) list of every Recipe belonging to the current user's Account.
      operationId: list-recipes
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/RecipeSummary"
                type:
                  - array
                  - "null"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List recipes
      tags:
        - Recipes
...
servers:
  - url: /api/bigshop